I did lot of upgrade for the rpg to-do list project. I was using modular build already but project directory were looking like a child made it. So I rearranged most of the files and also created a general_funcs.c file. This file will shorten my codes by giving functions that can be used on every plugin. So I didn't need to write them over and over again. Also this rearrangment is not something that I used to. So I prepared Makefile file with ai. Sorry. But don't worry! Non of those codes are written by ai. Only makefile. Anyway. I forgot all of my upgrades so commit message ends in here. Cya!

+++ plugin_src/tui_plugin/info_panel.c
@@ -21,7 +21,7 @@ int info_panel(){
 
 	if(width > 24){
 		int i = 1;
-		while(i < strlen(p1.expBar) - 1){
+		while(i < (int)strlen(p1.expBar) - 1){
 			if(p1.expBar[i] == '_') break;
 			if(p1.expBar[i] == '=') i++;
 		}
